dom/media/webaudio/test/test_channelMergerNode.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test ChannelMergerNode</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

var gTest = {
  length: 2048,
  numberOfChannels: 8,
  createGraph: function(context) {
    var buffers = [];
    for (var j = 0; j < 4; ++j) {
      var buffer = context.createBuffer(2, 2048, context.sampleRate);
      for (var i = 0; i < 2048; ++i) {
        buffer.getChannelData(0)[i] = Math.sin(440 * 2 * (j + 1) * Math.PI * i / context.sampleRate);
        // Second channel is silent
      }
      buffers.push(buffer);
    }
    var emptyBuffer = context.createBuffer(1, 2048, context.sampleRate);

    var merger = context.createChannelMerger();
    is(merger.channelCount, 2, "merger node has 2 input channels by default");
    is(merger.channelCountMode, "max", "Correct channelCountMode for the merger node");
    is(merger.channelInterpretation, "speakers", "Correct channelCountInterpretation for the merger node");

    for (var i = 0; i < 4; ++i) {
      var source = context.createBufferSource();
      source.buffer = buffers[i];
      source.connect(merger, 0, i);
      source.start(0);
    }
    return merger;
  },
  createExpectedBuffers: function(context) {
    var expectedBuffer = context.createBuffer(8, 2048, context.sampleRate);
    for (var i = 0; i < 4; ++i) {
      for (var j = 0; j < 2048; ++j) {
        expectedBuffer.getChannelData(i * 2)[j] = Math.sin(440 * 2 * (i + 1) * Math.PI * j / context.sampleRate);
      }
    }
    return expectedBuffer;
  },
};

runTest();

</script>
</pre>
</body>
</html>
